APAJAC-34: Testes unitários da Realização do Cars Testes unitários do serviço de cadastro do Cars e teste integrado do endpoint no controle de Cars.
