--- src/test/java/resources/application-test.properties
@@ -1,11 +1,0 @@
-spring.application.name=acolhimento
-server.port=8080
-spring.datasource.url=jdbc:h2:mem:public
-spring.datasource.driverClassName=org.h2.Driver
-spring.datasource.username=sa
-spring.datasource.password=password
-spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
-jpa.hibernate.ddl-auto=create
-spring.jpa.show-sql=true
-spring.jpa.defer-datasource-initialization=true
-spring.sql.init.mode=always
